+++ Classes/MMPReactiveNotification.m
@@ -63,6 +63,11 @@
     return self;
 }
 
+- (BOOL)notificationTypesUpToDate {
+    UIUserNotificationSettings *settings = [[UIApplication sharedApplication] currentUserNotificationSettings];
+    return settings.types == self.types;
+}
+
 #else
 
 - (instancetype)notificationTypes:(UIRemoteNotificationType)types {
@@ -114,12 +119,61 @@
     }];
 }
 
+- (RACSignal *)userNotificationSettingsRegistration {
+    return [[self.delegate rac_signalForSelector:@selector(application:didRegisterUserNotificationSettings:)
+                                    fromProtocol:@protocol(UIApplicationDelegate)]
+                           reduceEach:^id(id _, UIUserNotificationSettings *notificationSettings) {
+                               return notificationSettings;
+                           }];
+}
+
 - (RACSignal *)remoteNotifications {
     return [[self.delegate rac_signalForSelector:@selector(application:didReceiveRemoteNotification:)
                                     fromProtocol:@protocol(UIApplicationDelegate)]
                            reduceEach:^id(id _, NSDictionary *userInfo) {
                                return userInfo;
                            }];
+}
+
+- (RACSignal *)localNotifications {
+    return [[self.delegate rac_signalForSelector:@selector(application:didReceiveLocalNotification:)
+                                   fromProtocol:@protocol(UIApplicationDelegate)]
+                           reduceEach:^id(id _, UILocalNotification *notification) {
+                               return notification;
+                           }];
+}
+
+- (void)scheduleLocalNotification:(UILocalNotification *)notification {
+#if __IPHONE_OS_VERSION_MAX_ALLOWED >= 80000
+    if (![self notificationTypesUpToDate]) {
+        MMPRxN_LOG(@"Settings are not up-to-date, register settings before scheduling.")
+        [[[self userNotificationSettingsRegistration]
+                take:1]
+                subscribeNext:^(UIUserNotificationSettings *notificationSettings) {
+                    MMPRxN_LOG(@"Settings registered with types = %lu, now scheduling notification.", (unsigned long)notificationSettings.types)
+                    [[UIApplication sharedApplication] scheduleLocalNotification:notification];
+                }];
+        [[UIApplication sharedApplication] registerUserNotificationSettings:[UIUserNotificationSettings settingsForTypes:self.types
+                                                                                                              categories:self.categories]];
+    } else {
+        MMPRxN_LOG(@"Schedule notification immediately, settings are up-to-date.")
+        [[UIApplication sharedApplication] scheduleLocalNotification:notification];
+    }
+#else
+    [[UIApplication sharedApplication] scheduleLocalNotification:notification];
+#endif
+}
+
+- (void)scheduleLocalNotificationWithAlert:(NSString *)alertBody toBeFiredAt:(NSDate *)fireDate {
+    [self scheduleLocalNotificationWithAlert:alertBody withSound:UILocalNotificationDefaultSoundName toBeFiredAt:fireDate];
+}
+
+- (void)scheduleLocalNotificationWithAlert:(NSString *)alertBody withSound:(NSString *)soundName toBeFiredAt:(NSDate *)fireDate {
+    UILocalNotification *localNotification = [[UILocalNotification alloc] init];
+    localNotification.fireDate = fireDate;
+    localNotification.alertBody = alertBody;
+    localNotification.soundName = soundName;
+    [self scheduleLocalNotification:localNotification];
 }
 
 @end